requirements.txt
coverage==7.2.2
Django==4.1.7
django-environ==0.9.0
djangorestframework==3.14.0
djoser==2.2.0
environ==1.0
ofxparse==0.21
setuptools==65.7.0
simplejson==3.19.1
mysqlclient==2.1.1
pandas==1.5.3
pandera==0.14.3
pytest==7.2.2
sentry-sdk==1.18.0
